AoC D12 P2

+++ advent-of-code/2020/rust/day12/src/lib.rs
@@ -3,7 +3,7 @@ use std::str::FromStr;
 
 use anyhow::{anyhow, Error, Result};
 
-pub type Value = i16;
+pub type Value = i32;
 
 #[derive(Debug, Clone)]
 pub enum Action {
